phase2/backend/requirements.txt
fastapi==0.104.1
sqlmodel==0.0.16
uvicorn[standard]==0.24.0
pyjwt[crypto]==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
psycopg2-binary==2.9.9
alembic==1.13.1